requirements.txt
streamlit==1.45.1
pypdf==4.3.1
scikit-learn==1.5.1
groq==0.31.1
python-dotenv==1.0.1
